+++ internal/logic/analysis/character_analysis.go
@@ -5,7 +5,7 @@ import (
 	"sort"
 
 	"ba-torment-data-process/internal/constants"
-	"ba-torment-data-process/internal/logic"
+	"ba-torment-data-process/internal/logic/id"
 	"ba-torment-data-process/internal/types"
 )
 
@@ -60,7 +60,7 @@ func AnalyzeCharacter(studentID int, partyDataMap map[string]*types.BATormentPar
 		})
 
 		// starDistribution: group by groupID
-		groupID := ExtractGroupID(raidID)
+		groupID := id.ExtractGroupID(raidID)
 		if _, exists := groupStar[groupID]; !exists {
 			groupStar[groupID] = make(map[string]int)
 			groupOrder = append(groupOrder, groupID)
@@ -146,7 +146,7 @@ func analyzeCharacterInRaid(studentID int, partyData *types.BATormentPartyData) 
 					continue
 				}
 
-				memberStudentID, star, weaponStar, isAssist := logic.ParseStudentDetailID(member)
+				memberStudentID, star, weaponStar, isAssist := id.ParseStudentDetailID(member)
 
 				// Collect squad members for synergy
 				squadMembers = append(squadMembers, memberStudentID)
