+++ internal/logic/party/summary.go
@@ -1,4 +1,4 @@
-package parse
+package party
 
 import (
 	"fmt"
@@ -7,12 +7,11 @@ import (
 	"strings"
 
 	"ba-torment-data-process/internal/constants"
-	"ba-torment-data-process/internal/logic"
+	"ba-torment-data-process/internal/logic/id"
 	"ba-torment-data-process/internal/types"
 )
 
 func ProcessPartyDataToSummaryData(partyData *types.BATormentPartyData) (*types.BATormentSummaryData, error) {
-	// 토먼트/루나틱 데이터 분리
 	var lunaticData, tormentData []types.BATormentPartyDetail
 
 	isInsane := partyData.PartyDetail[0].Score < constants.TormentMinScore
@@ -33,7 +32,6 @@ func ProcessPartyDataToSummaryData(partyData *types.BATormentPartyData) (*types.
 	lunaticCount := len(lunaticData)
 	tormentCount := len(tormentData)
 
-	// 최종 데이터 구성
 	result := &types.BATormentSummaryData{
 		Torment: processLevelData(tormentData, "torment", lunaticCount, tormentCount),
 		Lunatic: processLevelData(lunaticData, "lunatic", lunaticCount, tormentCount),
@@ -42,7 +40,6 @@ func ProcessPartyDataToSummaryData(partyData *types.BATormentPartyData) (*types.
 	return result, nil
 }
 
-// processLevelData는 각 레벨의 데이터를 처리합니다.
 func processLevelData(data []types.BATormentPartyDetail, level string, lunaticCount, tormentCount int) types.BATormentLevelData {
 	result := types.BATormentLevelData{
 		ClearCount:  len(data),
@@ -52,7 +49,6 @@ func processLevelData(data []types.BATormentPartyDetail, level string, lunaticCo
 
 	parties := make([]string, 0)
 
-	// 필터 데이터 처리
 	filters := make(map[string](map[string]int))
 	assistFilters := make(map[string](map[string]int))
 
@@ -62,12 +58,12 @@ func processLevelData(data []types.BATormentPartyDetail, level string, lunaticCo
 			members := entry.PartyData[i]
 			var charIDs []string
 			for _, member := range members {
-				charID := logic.GetStudentID(member)
+				charID := id.GetStudentID(member)
 				if charID == 0 {
 					continue
 				}
 
-				logic.UpdateSummaryFilters(filters, assistFilters, member)
+				updateSummaryFilters(filters, assistFilters, member)
 				charIDs = append(charIDs, strconv.Itoa(charID))
 			}
 			sort.Strings(charIDs)
@@ -77,10 +73,8 @@ func processLevelData(data []types.BATormentPartyDetail, level string, lunaticCo
 		parties = append(parties, key)
 	}
 
-	// 파티 목록을 정렬
 	sort.Strings(parties)
 
-	// 빈 배열이면 조기 종료
 	if len(parties) == 0 {
 		return result
 	}
@@ -94,7 +88,6 @@ func processLevelData(data []types.BATormentPartyDetail, level string, lunaticCo
 	currentKey := parties[0]
 	currentCount := 1
 
-	// 정렬된 목록에서 연속된 같은 값의 개수를 세어 저장
 	for i := 1; i < len(parties); i++ {
 		if parties[i] == currentKey {
 			currentCount++
@@ -106,14 +99,13 @@ func processLevelData(data []types.BATormentPartyDetail, level string, lunaticCo
 	}
 	usages = append(usages, partyUsage{currentKey, currentCount})
 
-	// 사용 횟수로 정렬하고, 같은 횟수면 키로 정렬
 	sort.Slice(usages, func(i, j int) bool {
 		if usages[i].count != usages[j].count {
 			return usages[i].count > usages[j].count
 		}
 		return usages[i].key < usages[j].key
 	})
-	// Top 5 파티 추출
+
 	for i := 0; i < 5 && i < len(usages); i++ {
 		party := []any{
 			usages[i].key,
@@ -122,10 +114,8 @@ func processLevelData(data []types.BATormentPartyDetail, level string, lunaticCo
 		result.Top5Partys = append(result.Top5Partys, party)
 	}
 
-	// 임계값에 따른 파티 카운트 계산
 	thresholds := getThresholds(lunaticCount, tormentCount, level == "torment")
 	for _, threshold := range thresholds {
-		// 파티 수에 따른 카운트 계산 [1파티, 2파티, 3파티, 4파티 이상]
 		partyCounts := make([]int, 4)
 		for i := range data {
 			entry := data[i]
@@ -134,7 +124,6 @@ func processLevelData(data []types.BATormentPartyDetail, level string, lunaticCo
 				continue
 			}
 			partyData := entry.PartyData
-			// 파티 데이터의 키 개수로 파티 수 계산
 			numParties := len(partyData)
 			if numParties >= 4 {
 				partyCounts[3]++
@@ -148,7 +137,6 @@ func processLevelData(data []types.BATormentPartyDetail, level string, lunaticCo
 	return result
 }
 
-// getThresholds는 클리어 수에 따라 적절한 임계값 목록을 반환합니다.
 func getThresholds(lunaticCount, tormentCount int, isTorment bool) []int {
 	allThresholds := []int{100, 200, 500, 1000, 2000, 5000, 10000, 20000}
 	var thresholds []int
